+++ database/migrations/2020_03_29_171656_create_salary_logs_table.php
@@ -20,7 +20,7 @@ class CreateSalaryLogsTable extends Migration
             $table->decimal('attendence_number', 8, 2)->default(0);
             $table->integer('food_rate_will_get');
             $table->integer('food_rate_paid')->default(0);
-            $table->bigInteger('group_id');
+            $table->integer('group_id');
             $table->bigInteger('building_id');
 
             $table->timestamps();
